requirements.txt
annotated-types==0.6.0
anyio==4.3.0
bson==0.5.10
certifi==2024.2.2
click==8.1.7
cloudinary==1.39.1
dnspython==2.6.1
docutils==0.21
fastapi==0.110.1
h11==0.14.0
httptools==0.6.1
idna==3.6
numpy==1.26.4
opencv-python==4.9.0.80
pillow==10.3.0
pydantic==2.6.4
pydantic_core==2.16.3
pymongo==4.6.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
